@jcorbin 2016-01-29
:ok: with me since the tests green, nice benchmark numbers